refactor(quality): simplify adapter verification units
Constraint: Preserve every adapter contract and sandbox observation while keeping each source and test unit below 250 pure lines.
Rejected: Leave the 407-line contract test intact | it violated the user-requested reviewable-unit boundary.
Confidence: high
Scope-risk: narrow
Directive: Add future corpus projection cases to the dedicated projection test module, not the core schema contract module.
Tested: 36 focused, 99 related, 2982 full tests; contract probe; actual macOS sandbox probes; Ruff; compileall; no-excuse; diff and privacy checks.
Not-tested: basedpyright was unavailable.
